requirements.txt
fastapi
uvicorn
kafka-python
elasticsearch
requests